+++ Sources/BaseFoundationiOS/Protocols/CoreProtocols/FirstCaseDataSource.swift
@@ -1,8 +1,12 @@
 //
-//  File.swift
+//  FirstCaseDataSource.swift
 //  
 //
 //  Created by pablo borquez on 06/08/23.
 //
 
 import Foundation
+
+public protocol FirstCaseDataSource {
+    static var firstCase: Self { get }
+}