+++ Sources/BaseFoundationiOS/Templates/Types/PickerToolbarType.swift
@@ -1,0 +1,10 @@
+//
+//  PickerToolbarType.swift
+//  
+//
+//  Created by pablo borquez on 06/08/23.
+//
+
+import Foundation
+
+public typealias PickerToolbarType = CaseIterable & Identifiable & Hashable & ValueDataSource & FirstCaseDataSource